+++ pages/reset-password/index.vue
@@ -34,7 +34,7 @@ async function onSubmit() {
   success.value = false;
   loading.value = true;
 
-  const { error: sendError } = await tryCatch(resetPassword(email.value));
+  const { error: sendError } = await tryRequestEndpoint(`/auth/password/reset/`, "POST", { email: email.value });
 
   loading.value = false;
