+++ utils/types.ts
@@ -1,119 +1,3 @@
-export interface Answer {
-  /** @readonly ID of the answer. */
-  readonly id: number;
-  /** @readonly What the answer choice says (HTML string). */
-  readonly text: string;
-  /** Used to store which answer the student selected.
-   * @warning Must be manually added to `Answer`; this field is not returned from the API.
-   */
-  selected: boolean;
-  /** Whether or not the answer is correct.
-   * @warning Must be added manually. Should be added after submitting a question and receiving results.
-   */
-  isCorrect?: boolean;
-}
-
-export interface Question {
-  /** @readonly What the question says (HTML string). */
-  readonly text: string;
-  /** @readonly The type of answer for the question. */
-  readonly answerType: "Multiple Choice" | "Written Response" | "True or False";
-  /** @readonly Array of answers. */
-  readonly answers: Answer[];
-}
-
-interface QuestionInterface {
-  /** @readonly ID of the question interface. */
-  readonly id: number;
-  question: Question;
-}
-
-export interface DynamicQuestionInterface extends QuestionInterface {
-  /** Number of attempts used on this dynamic question. */
-  answerAttemptsUsed: number;
-}
-
-export interface StaticQuestionInterface extends QuestionInterface {
-  /** Index of the question.
-   * @warning Only present if questionIndex was not passed in the URL.
-   * @warning Starts at 1, with 0 indicating an unknown index.
-   */
-  questionIndex: number;
-  staticUserAnswer: number | null;
-}
-
-export interface TopicQuestionInterfaceAnswer {
-  id: number;
-  /** @readonly What the answer choice says (HTML string). */
-  text: string;
-  isCorrect: boolean;
-}
-export interface TopicQuestionInterface {
-  id: number;
-  /** @readonly What the answer choice says (HTML string). */
-  readonly text: string;
-  answerType: "Multiple Choice" | "Written Response" | "True or False";
-  difficulty: number;
-  answers: TopicQuestionInterfaceAnswer[];
-  correctFirstAttempts: number;
-  totalFirstAttempts: number;
-}
-
-interface TeacherAssignmentStatisticData {
-  /** ID of the assignment instance */
-  assignmentInstance: number;
-  /** ID of the question */
-  question: number;
-  /** Time spent on the assignment, in seconds */
-  timeSpent: number;
-}
-/** @template T - Whether the `guaranteedQuestions` field should be an array of `Question` objects or an array of question IDs */
-interface TeacherAssignmentStatistic<T extends boolean> {
-  statisticsData: TeacherAssignmentStatisticData;
-  /** Array of guaranteed questions if `T` is true, question IDs if false */
-  guaranteedQuestions: T extends true ? Question[] : number[];
-}
-export interface DynamicTeacherAssignmentStatistic<T extends boolean = true> extends TeacherAssignmentStatistic<T> {
-  statisticsData: TeacherAssignmentStatisticData & {
-    /** Array of answer IDs that represent the user's answers for the question */
-    dynamicUserAnswers: number[];
-  };
-}
-export interface StaticTeacherAssignmentStatistic<T extends boolean = true> extends TeacherAssignmentStatistic<T> {
-  statisticsData: TeacherAssignmentStatisticData & {
-    /** Answer ID */
-    staticUserAnswer: number;
-  };
-}
-
-interface IndividualStudentStatistic<T extends boolean> {
-  /** ID of the assignment instance */
-  assignmentInstance: number;
-  /** Measured in seconds */
-  timeSpent: number;
-  /** Array of questions if `T` is true, question IDs if false */
-  question: T extends true ? TopicQuestionInterface : number;
-}
-export interface DynamicIndividualStudentStatistic<T extends boolean = true> extends IndividualStudentStatistic<T> {
-  /** Array of answer IDs that represent the user's answers for the question */
-  dynamicUserAnswers: number[];
-}
-export interface StaticIndividualStudentStatistic<T extends boolean = true> extends IndividualStudentStatistic<T> {
-  staticUserAnswer: number;
-}
-
-export interface Topic {
-  id: number;
-  name: string;
-  hasChildren: boolean;
-  hasQuestions: boolean;
-}
-export interface TopicMapped extends Topic {
-  /** IDs of child topics */
-  children: number[] | null;
-  questionIds: number[];
-}
-
 export interface CreateCourse {
   id: number;
   joinCode: string;
@@ -145,128 +29,6 @@ export interface StudentStatistic<T extends boolean = true> {
   timeStarted: Date | null;
 }
 
-interface Assignment {
-  /** @readonly ID of the question. */
-  readonly id: number;
-}
-
-export interface StudentAssignment extends Assignment {
-  /**Date object of when the assignment was submitted. */
-  dateSubmitted: Date | null;
-
-  /** Number of questions completed. */
-  questionsCompleted: number;
-
-  /** Number of correct questions in the assignment */
-  questionsCorrect: number;
-
-  /** @readonly assignment object for assignment properties. */
-
-  readonly assignment: {
-    /** @readonly Number of attempts allowed, if assignment is dynamic */
-    readonly attemptsAllowed: number;
-
-    /** @readonly Name of the assignment. */
-    readonly name: string;
-
-    /** @readonly Number of questions in the assignment */
-    readonly numQuestions: number;
-
-    /** @readonly Whether or not the assignment can be turned in late. */
-    readonly lateSubmissions: boolean;
-
-    /**Date object of when the assignment is due (Date(UTC)). */
-    dueDate: Date;
-
-    /**Date object of when the assignment was assigned (Date(UTC)). */
-    dateAssigned: Date;
-
-    /** @readonly If the assignment is a static assignment.*/
-    readonly isStatic: boolean;
-
-    /** @readonly Object identifying the course assignment belongs to.
-     * @warning Only present if assignment is fetched for all courses.
-     */
-    readonly course?: {
-      /** @readonly Id of the course assignment belongs to */
-      readonly id: number;
-      /** @readonly Name of the course assignment belongs to */
-      readonly name: string;
-      readonly subject: Subject;
-    };
-
-    /**
-     * Used to store already-fetched `QuestionInterface`s for easy access when going back and forth.
-     *
-     * Each key is the index of the question in the assignment.
-     * @warning Must be manually added to `StudentAssignment`; this field is not returned from the API.
-     * @warning **Questions at certain indices may not exist yet**; the question must be fetched from the API first before being added.
-     */
-    questionInterfaces: Record<number, QuestionInterface>;
-  };
-}
-
-export interface TeacherAssignment extends Assignment {
-  /** @readonly Name of the assignment. */
-  readonly name: string;
-
-  /** Date the assignment was submitted (Date(UTC)) */
-  dateAssigned: Date;
-
-  /** Date the assignment is due (Date(UTC)) */
-  dueDate: Date;
-
-  /** @readonly Number of questions in the assignment */
-  readonly numQuestions: number;
-
-  /** @readonly Whether or not the assignment can be turned in late. */
-  readonly lateSubmissions: boolean;
-
-  /** @readonly Number of questions submitted */
-  readonly numSubmitted: number;
-}
-
-interface Course {
-  /** @readonly ID of the course. */
-  readonly id: number;
-  /** @readonly Name of the course. */
-  readonly name: string;
-  /** @readonly Name of the teacher in First Name Last Name format.
-   * @example "Michael Whalen"
-   */
-  readonly teacher: string;
-  /** @readonly Period of the course. */
-  readonly period: number;
-  /** @readonly Subject of the course. */
-  readonly subject: Subject;
-  /** @warning This field is not returned from the API, must be added manually on fetch */
-  assignmentsFetched: boolean;
-}
-
-export interface StudentCourse extends Course {
-  assignments: StudentAssignment[];
-}
-
-export interface TeacherCourseNoAssignment extends Course {
-  /** @warning This field is not returned from the API, must be added manually on fetch */
-  assignments: TeacherAssignment[];
-  /** @readonly 6-digit join code for the course. */
-  readonly joinCode: string;
-  /** The number of students in the course. */
-  numStudents: number;
-}
-
-export interface TeacherCourse extends TeacherCourseNoAssignment {
-  assignments: TeacherAssignment[];
-}
-
-export interface AssignmentInstance {
-  /** @readonly ID of the assignment instance. */
-  readonly assignment: number;
-  /** @readonly ID of the student. */
-  readonly student: number;
-}
-
 export interface SubmitAnswer {
   /** @readonly Tells you if the answer was correct. */
   readonly isCorrect: boolean;
@@ -274,34 +36,11 @@ export interface SubmitAnswer {
   readonly remainingAttempts: number | null;
 }
 
-export interface SubmitAssignment {
-  /** @readonly Number of questions in the assignment. */
-  readonly numQuestions: number;
-  /** Number of questions completed. */
-  questionsCompleted: number;
-  /** Number of questions correct. */
-  questionsCorrect: number;
-}
-
-export interface AssignmentResults extends SubmitAssignment {
-  /** @readonly An array of the question results. */
-  readonly questionInstances: {
-    /** @readonly ID of the question. */
-    readonly id: number;
-    /** @readonly The data for the question referenced by the instance. */
-    readonly question: Question;
-    /** @readonly Array of dynamic user answers (IDs of selected answers). */
-    readonly dynamicUserAnswers: number[] | null;
-    /** @readonly The ID of the chosen answer if the assignment is static.*/
-    readonly staticUserAnswer: number | null;
-    /** @readonly Tells if the question is complete. */
-    readonly isComplete: boolean;
-    /** @readonly Time spent on the question (in seconds). */
-    readonly timeSpent: number;
-  }[];
-}
-
-export interface CreateAssignmentQuestion {
-  questionId: number;
-  isGuaranteed: boolean;
-}
+// https://nuxt.com/docs/guide/directory-structure/composables#how-files-are-scanned
+export * from "./types/teacherStatistics";
+export * from "./types/assignments";
+export * from "./types/questions";
+export * from "./types/subjects";
+export * from "./types/courses";
+export * from "./types/topics";
+export * from "./types/todo";
